Make language settings safer against uninstallation of this mod.
It was already possible to uninstall this mod and continue to play
the game, but entering the options would crash, unless another thing
in the game saved something.

Now, uninstalling this mod should be entirely safe, since we always
save the english language in the option when using an added locale,
and we ignore it on load.

The only thing that saves the current language is now localStorage's
IG_LANG, and the game handles unknown IG_LANG properly, by resetting it
to english.