Workaround a bug of CryptJS when failing to decrypt.
It seems that this version of CryptoJS is even more broken than
anticipated. When the decryption of something invalid returns a string
of length 32 and the last byte is e.g. 232, then that idiot,
instead of indicating that the PKCS#7 padding failed, would
instead return a Words of length -200, and then would fail horribly
at handling that.

It could have returned something else, i don't know, null ?

So check this horrible case, and wrap all code touching CryptoJS into
a pokemon-like exception handler.