Add option `patch_base_font`, which does not add any feature.
It was a bad decision to have `patch_font` be responsible for
patching both the base font and the colors.  That's stupid.

Now, if `patch_base_font` is defined, it is called first to patch the base
font, and `patch_font` is used for the colors images.

And if `patch_base_font` is defined but `patch_font` isn't, then
by default, colors images will be made from the base font, as
if context.recolor_from_base_image(canvas) was called.  This means
mods don't even need to worry about colors anymore.