Do not block loading files which do not have lang labels.
Some mods block until their JSON resources are loaded. This creates a
deadlock because Localize-Me want to translate the strings in the JSON,
but to do that, it blocks until the final locale is known, which can
only happen once the game starts.

As a workaround, make Localize-Me not block resources which do not have
any lang labels.

If you want your mod to be translatable, do not block in your postload.
If the game is able to not block on a resource during initialization, so
can you.

This patch seems to improves the game initial loading time.