Add support for translating stuff inside extension/
Previously, we only searched for stuff to translate inside assets/data.
But since assets/extensions may contain the postgame DLC and/or the
description of the ninja skin, stuff inside the extension/ directory
probably need to be translated too.

The file_path for stuff inside extensions currently starts with
"extension/", as if the content was in "assets/data/extension/",
e.g. "extension/post-game-dlc/post-game-dlc.json".

This still requires a documentation update.