Fix crash when a modded locale is set, but the save file doesn't exist or is empty
[lesser.sherry: Cherry-picked from the v1.x branch

This can only happen if currentLang is an added locale and there is no
save file.  The easiest way for this to happen is to have an added
locale that defines useFor and that locale gets selected by the game as
a result.  Otherwise, currentLang will default to a builtin locale and
the crash will not happen.]