Add support for mod-provided langfile translations.
If a mod add fields inside a langfile, then the mod couldn't provide its
own translation for added locales.  The only option was for the
translation mod to handle it.

This patch adds the possibility for the mod to have its own translation
by simply adding a assets/lang/sc/stuff.tr_TR.json.patch to their
mod, even for added locales.

This previously was not possible for added locales, because
Localize-Me would redirect stuff.tr_TR.json to stuff.en_US.json (if
from_locale is en_US), and this would be oblivious to whatever
modloader implementation out there.

This is done by using ccloader v3's asset generation feature, and
letting it patch the generated langfile.