Fix indentation and add back the license field.
ccmod.json was initially implemented by ccloader3.  It used to have no
specification (as usual) so the only point of reference was other
ccloader3 mods like crosscode-ru and the ccloader3 source code.

ccloader2 eventually supported ccmod.json.  Lately, a new ccmod.json
specification is being written.  As of 2024-03-04¹ it is quite vague and
is missing the license field (despite being widely used in ccloader3
mods) but it adds the "tags" field.

The "authors" field is currently not specified as a copyright
statement, but that does not mean that the Berne convention does not
apply, so keep the license field for now.

Let's hope that the ccloader2 and ccloader3 expectation of ccmod.json
do not conflict in the future, because this mod is a dependency of both
ccloader2 and ccloader3 mods, as well as mods that supports both.

[1]: https://github.com/CCDirectLink/CCModDB/blob/39e0db315c72869c88ebb2767fcc2cefd1123ebc/CCMOD-STANDARD.md